# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Symfony E-commerce API",
        "description": "API REST do e-commerce. Para endpoints protegidos, fa\u00e7a login em /api/login e use o token no bot\u00e3o \"Authorize\".",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/products/{id}": {
            "get": {
                "tags": [
                    "Produtos (admin)"
                ],
                "summary": "Retorna os detalhes de um produto, inclusive arquivado (somente administradores).",
                "operationId": "get_api_admin_products_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-[13-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "N\u00e3o autenticado"
                    },
                    "403": {
                        "description": "Sem permiss\u00e3o de administrador"
                    },
                    "404": {
                        "description": "Produto n\u00e3o encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductResponse": {
                "required": [
                    "id",
                    "sku",
                    "name",
                    "slug",
                    "priceInCents",
                    "active",
                    "createdAt",
                    "updatedAt"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "01a0e4a1-a692-7c53-9cc0-add061fe64e8"
                    },
                    "sku": {
                        "type": "string",
                        "example": "CAM-001"
                    },
                    "name": {
                        "type": "string",
                        "example": "Camiseta B\u00e1sica Preta"
                    },
                    "slug": {
                        "type": "string",
                        "example": "camiseta-basica-preta-cam-001"
                    },
                    "description": {
                        "type": "string",
                        "example": "Camiseta 100% algod\u00e3o",
                        "nullable": true
                    },
                    "priceInCents": {
                        "description": "Pre\u00e7o em centavos",
                        "type": "integer",
                        "example": 4990
                    },
                    "active": {
                        "type": "boolean"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ],
...
}